reorganise and add comments
