+++ src/DetectorImage.jsx
@@ -1,5 +1,20 @@
 import React, {useState} from "react";
 
+/**
+ * Component that displays an image processed by a specific detector
+ * 
+ * @component
+ * @author Samuel Niang
+ * 
+ * @param {Object} props - Component props
+ * @param {string} props.selectedFile - The name of the image file to display
+ * @param {string} props.selectedDetector - The type of detector used to process the image
+ * 
+ * @returns {JSX.Element} A div containing the processed image
+ * 
+ * @example
+ * <DetectorImage selectedFile={selectedFile} selectedDetector={selectedDetector} />
+ */
 const DetectorImage = ({selectedFile, selectedDetector}) => {
     return (
         <div>
